check file permissions before deleting
This patch ensures that we have the correct permissions to edit/delete a
file before we attempt to delete it. If we lack the correct permissions
then we will simply skip deleting it.

We can find ourselves in this situation when the user has run vim with
superuser privileges (sudo vim). This will cause vim to write the backup
files as root. For now we will simply skip deleting such files. If the
user wants such files clean out, then running vim as a superuser will
give central.vim the correct privileges to do so.