fix bug: vim dir created wherever vim is run from
The `$VIMHOME` variable is not defined by Vim itself, instead it is
something that I have defined in my personal vimrc file. This patch
resolves this issue by defining `$VIMHOME` within this plugin. Note
that it will not redefine `$VIMHOME` if it already exists; allowing
the user to manually configure their own location to store the back
up files if they so wish.